App class, City class, and State class Added
The City and State classes were defined for later use in the App class. The ability to add, delete, and list all the states from the user was added. The ability to add and delete a city from a state was also added. The framework to list all the cities of a state was added.